pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Anastasiia Ermakova <ermakova.9999@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.3"
django-bootstrap-datepicker-plus = "^5.0.0"
djangorestframework = "3.14.0"
Pillow = "^9.3.0"
django-crispy-forms = "^1.14.0"
pytz = "^2022.6"
psycopg2-binary = "^2.9.5"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
